Refactor: FooterTopLayout 컴포넌트의 이름을 FooterTopContainer로 변경 - 프로젝트의 파일 구조에서, Page 컴포넌트의 자식 컴포넌트로 Container를 생성하기 때문에 그에 맞게 파일명을 변경
