+++ src/Footer/FooterTopContainer.test.jsx
@@ -1,22 +1,22 @@
 import { render } from '@testing-library/react';
 
-import FooterTopLayout from './FooterTopLayout';
+import FooterTopContainer from './FooterTopContainer';
 
 describe('FooterTopContainer', () => {
   it('renders the title', () => {
-    const { container } = render(<FooterTopLayout />);
+    const { container } = render(<FooterTopContainer />);
 
     expect(container).toHaveTextContent('Shopify');
   });
 
   it('renders the list', () => {
-    const { container } = render(<FooterTopLayout />);
+    const { container } = render(<FooterTopContainer />);
 
     expect(container).toHaveTextContent('Privacy Policy');
   });
 
   it('renders the info', () => {
-    const { container } = render(<FooterTopLayout />);
+    const { container } = render(<FooterTopContainer />);
 
     expect(container).toHaveTextContent('GitHub');
   });
